issue 16644 - bluestore asserts when block device not an even 4096 multiple
BitmapFreelistManager::create() already checks to see if the block device is
an even multiple of (bytes_per_block * blocks_per_key), and it updates the
bitmap to make any excess bytes at the top of the device appear to be in use.
However, if the block device is not an even multiple of 4096 bytes, bluestore
asserts.  The code needs to be adjusted to include any partial-4096 block
within the byte range that it marks as being in use.

Fixes: #16644
Signed-off-by:krehm